backend/requirements.txt
amqp==2.5.2
asgiref==3.2.5
astroid==2.3.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.3
django-cachalot==2.2.0
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
dnspython==1.16.0
eventlet==0.25.2
googletrans==3.0.0
greenlet==0.4.15
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.25
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.9
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
Markdown==3.3.2
mccabe==0.6.1
monotonic==1.5
mysqlclient==1.4.6
Pillow==7.1.1
pylint==2.4.4
pytimeparse==1.1.8
pytz==2019.3
redis==3.5.3
requests==2.23.0
rfc3986==1.4.0
six==1.14.0
sniffio==1.1.0
soupsieve==2.0.1
sqlparse==0.3.0
urllib3==1.25.9
vine==1.3.0
wincertstore==0.2
wrapt==1.12.1
